More work on hubbard

+++ src/hubbard/hubbard_potential_energy.cpp
@@ -608,7 +608,7 @@ one_electron_energy_hubbard(Hubbard_matrix const& om__, Hubbard_matrix const& pm
                 auto src1 = om__.local(at_lvl).at(memory_t::host);
                 auto src2 = pm__.local(at_lvl).at(memory_t::host);
 
-                for (int i = 0; i < om__.local(at_lvl).size(); i++) {
+                for (int i = 0; i < static_cast<int>(om__.local(at_lvl).size()); i++) {
                     tmp += src1[i] * std::conj(src2[i]);
                 }
             }
